# backend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "fulkruma",
  "version": "0.0.1"
 },
 "paths": {
...
  "/api/v1/webhooks/endpoints": {
   "get": {
    "operationId": "getWebhooksEndpoints",
    "summary": "List endpoints",
    "tags": [
     "webhooks"
    ],
    "parameters": [],
    "x-forjio": {
     "source": "src/routes/webhooks.ts:64",
     "guards": [
      "requireAuth"
     ]
    }
   },
   "post": {
    "operationId": "postWebhooksEndpoints",
    "summary": "Register an endpoint.",
    "description": "Register an endpoint. The response is the only time its signing secret\nis returned. The URL must be https (in production) and must not point\nat a private, loopback or link-local address. `events` takes \"*\",\nevent types (\"fulkruma.shipment.created.v1\") and prefixes\n(\"fulkruma.shipment.*\").",
    "tags": [
     "webhooks"
    ],
    "parameters": [],
    "x-forjio": {
     "source": "src/routes/webhooks.ts:82",
     "guards": [
      "requireAuth"
     ],
     "body": "validated",
     "bodySchema": "createSchema"
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "properties": {
         "url": {
          "type": "string",
          "format": "uri",
          "maxLength": 2000
         },
         "events": {
          "type": "array",
          "items": {
           "type": "string",
           "minLength": 1,
           "maxLength": 200
          },
          "minItems": 1,
          "maxItems": 50,
          "default": [
           "*"
          ]
         },
         "description": {
          "type": "string",
          "maxLength": 500
         }
        },
        "required": [
         "url"
        ],
        "additionalProperties": false
       }
      }
     }
    }
   }
  },
  "/api/v1/webhooks/endpoints/{id}": {
   "patch": {
    "operationId": "patchWebhooksEndpointsId",
    "summary": "Update an endpoint.",
    "description": "Update an endpoint. Setting `active: false` pauses it (its queued\ndeliveries become failed); `active: true` re-enables it — also after\nFulkruma switched it off for failing — and clears its failure streak.\nA new `url` goes through the same checks as on register; the signing\nsecret stays the same.",
    "tags": [
     "webhooks"
    ],
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "x-forjio": {
     "source": "src/routes/webhooks.ts:113",
     "guards": [
      "requireAuth"
     ],
     "body": "validated",
     "bodySchema": "updateSchema"
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "properties": {
         "url": {
          "type": "string",
          "format": "uri",
          "maxLength": 2000
         },
         "events": {
          "type": "array",
          "items": {
           "type": "string",
           "minLength": 1,
           "maxLength": 200
          },
          "minItems": 1,
          "maxItems": 50
         },
         "description": {
          "type": "string",
          "maxLength": 500
         },
         "active": {
          "type": "boolean"
         }
        },
        "additionalProperties": false
       }
      }
     }
    }
   },
   "delete": {
    "operationId": "deleteWebhooksEndpointsId",
    "summary": "Delete an endpoint",
    "tags": [
     "webhooks"
    ],
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "x-forjio": {
     "source": "src/routes/webhooks.ts:144",
     "guards": [
      "requireAuth"
     ]
    }
   }
  },
  "/api/v1/webhooks/events": {
   "get": {
    "operationId": "getWebhooksEvents",
    "summary": "List webhook deliveries.",
    "description": "List webhook deliveries. Newest first: one row per event per endpoint,\nwith its status (pending, sent, failed), attempt count, next retry and\nevery attempt made (`deliveryAttempts`). Filter by `type`, `status` or\n`endpointId`; page with `limit` (1-200, default 50) and the returned\n`nextCursor`.",
    "tags": [
     "webhooks"
    ],
    "parameters": [
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "maximum": 200,
       "default": 50
      }
     },
     {
      "name": "cursor",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "minLength": 1
      }
     },
     {
      "name": "type",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "minLength": 1
      }
     },
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "enum": [
        "pending",
        "sent",
        "failed"
       ]
      }
     },
     {
      "name": "endpointId",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "minLength": 1
      }
     }
    ],
    "x-forjio": {
     "source": "src/routes/webhooks.ts:167",
     "guards": [
      "requireAuth"
     ]
    }
   }
  },
  "/api/v1/webhooks/events/{id}": {
   "get": {
    "operationId": "getWebhooksEventsId",
    "summary": "Get a webhook delivery.",
    "description": "Get a webhook delivery. Includes every attempt made at it.",
    "tags": [
     "webhooks"
    ],
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "x-forjio": {
     "source": "src/routes/webhooks.ts:197",
     "guards": [
      "requireAuth"
     ]
    }
   }
  },
  "/api/v1/webhooks/events/{id}/retry": {
   "post": {
    "operationId": "postWebhooksEventsIdRetry",
    "summary": "Retry a webhook delivery.",
    "description": "Retry a webhook delivery. Queues one more attempt now at a failed\ndelivery (or sends a sent one again). It goes out within a few\nseconds; read it back with GET /webhooks/events/:id. 409 when it is\nalready queued or its endpoint is disabled.",
    "tags": [
     "webhooks"
    ],
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "x-forjio": {
     "source": "src/routes/webhooks.ts:212",
     "guards": [
      "requireAuth"
     ],
     "body": "none"
    }
   }
  },
...
 },
...
}